Move upstream inventory utils to separate file
