+++ common/go-utils/src/bin/update_inventory.rs
@@ -3,7 +3,7 @@
 
 use heroku_go_utils::vrs::GoVersion;
 use heroku_inventory_utils::inv::Inventory;
-use heroku_inventory_utils::inv::UpstreamInventory;
+use heroku_inventory_utils::upstream::UpstreamInventory;
 use std::{env, process};
 
 /// Updates the local go inventory.toml with versions published on go.dev.
