+++ common/inventory-utils/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod checksum;
 pub mod inv;
 pub mod semvrs;
+pub mod upstream;
 pub mod vrs;
